clang again

+++ code/scripting/api/objs/order.cpp
@@ -290,7 +290,7 @@ ADE_VIRTVAR(Target, l_Order, "object", "Target of the order. Value may also be a
 		case AI_GOAL_WAYPOINTS:
 		case AI_GOAL_WAYPOINTS_ONCE:
 			// check if waypoint order is the current goal (ohp->odx == 0) and if it is valid
-			if ( (ohp->odx == 0) && (aip->wp_index != INVALID_WAYPOINT_POSITION) && (aip->wp_index >= 0) &&
+			if ( (ohp->odx == 0) && (aip->wp_index != INVALID_WAYPOINT_POSITION) && (aip->wp_index >= static_cast<size_t>(0)) &&
 				(aip->wp_index < aip->wp_list->get_waypoints().size()) ) {
 				objnum = aip->wp_list->get_waypoints()[aip->wp_index].get_objnum();
 			} else {
